add alert in ruttien page

+++ app/manaTransaction/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 # Create your models here.
 class InfoUser(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    balance = models.IntegerField(default=0)
+    balance = models.FloatField(default=0)
     
     def __str__(self):
         return self.user.username
